+++ src/node/selector.rs
@@ -157,10 +157,10 @@ impl selectors::Element for NodeRef<'_> {
 
     /// Whether this element is a `link`.
     fn is_link(&self) -> bool {
-        // TODO: This function adds some overhead.
+        // This function adds some overhead.
         // Its purpose in dom_query is unclear.
         // Returning `false` works just fine.
-        self.query_or(false, |node| node.is_link())
+        false
     }
 
     /// Whether the element is an HTML element.