+++ src/mini_selector.rs
@@ -2,5 +2,5 @@ mod extension;
 mod parser;
 mod selector;
 
-pub use parser::{parse_selector_list, parse_mini_selector};
+pub use parser::{parse_mini_selector, parse_selector_list};
 pub use selector::MiniSelector;