+++ tests/selection-traversal.rs
@@ -567,3 +567,38 @@ fn test_selection_id() {
     assert_eq!(doc.select("body").id(), None);
     assert_eq!(doc.select("#non-existing").id(), None);
 }
+
+#[cfg_attr(not(target_arch = "wasm32"), test)]
+#[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
+fn test_selection_is_sorted() {
+    let doc: Document = ANCESTORS_CONTENTS.into();
+
+    let sel_0 = doc.select("div");
+    let nodes_id_0 = sel_0.nodes().iter().map(|n| n.id).collect::<Vec<_>>();
+    assert!(nodes_id_0.is_sorted());
+
+    let sel_1 = doc.select("#great-ancestor > div").select("div > div");
+    let nodes_id_1 = sel_1.nodes().iter().map(|n| n.id).collect::<Vec<_>>();
+    assert!(nodes_id_1.is_sorted());
+
+    let sel_2 = doc.select("#great-ancestor div").select("div > div");
+    let nodes_id_2 = sel_2.nodes().iter().map(|n| n.id).collect::<Vec<_>>();
+    assert!(nodes_id_2.is_sorted());
+}
+
+
+#[cfg_attr(not(target_arch = "wasm32"), test)]
+#[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
+fn test_select_single_ancestors() {
+    let doc: Document = ANCESTORS_CONTENTS.into();
+
+    let nonexisting_sel = doc.select("#ancestor").select("#parent").select_single("div");
+    assert!(!nonexisting_sel.exists());
+
+    let div_sel = doc.select_single("#great-ancestor").select_single("div");
+    assert!(div_sel.exists());
+
+    let p_sel = doc.select_single("#great-ancestor").select_single("p");
+    assert!(!p_sel.exists());
+    
+}